+++ app/controllers/comments_controller.rb
@@ -4,8 +4,8 @@ class CommentsController < ApplicationController
     commentable = Comment.find_commentable params[:comment][:commentable_type], params[:comment][:commentable_id]
 
     @comment = commentable.comments.new(params[:comment])
-    @comment.user_id = current_user.id if !current_user.nil?
-    @comment.approved = !commentable.respond_to?("moderated") || !commentable.moderated || commentable.owner == current_user
+    @comment.user_id = current_user.id unless current_user.nil?
+    @comment.approved = !commentable.respond_to?("moderated") || !commentable.moderated || commentable.commentable_owner == current_user
 
     @comment.spam = Cerberus.check_spam(@comment, request)